Render deprecation metadata in all help formats
docstring_parser strips a top-level `.. deprecated::` regardless of
help_format, and the default format is markdown - so gating the
reconstruction on rst left the fix inert for default apps. The tag text
renders literally in every format, so drop the gate.

Move format_deprecated_tag into help.py: rst_preprocessor.py is deleted
on v5-develop, and this placement lets the same patch apply to both
branches. Command-list entries show the tag only; the full deprecation
message renders on the command's own help page.
